requirements.txt
Django==1.10.2
djangorestframework==3.1.3
requests==2.11.1
django-allauth==0.28.0
factory-boy==2.8.1
